Add files via upload
This is a Javascript function that finds and returns the longest word in the sentence with multiple lowercase words separated by spaces. If there are multiple words of the same length, it chooses the one that has the highest number of vowels. (Ignoring any characters in the sentence that is not an English letter or a space.)


Sample input:


“Smart people learn from everything and everyone, average people from their experience, stupid people already, have all the answers” (Socrates)